+++ lib/down/errors.rb
@@ -17,7 +17,7 @@ module Down
   class ResponseError < Error
     attr_reader :response
 
-    def initialize(message, response: nil)
+    def initialize(message, response = nil)
       super(message)
       @response = response
     end